refactor: useQuery 적용 및 fsd 폴더 구조 일부 적용
